Minor changes
Changed some scan options as they were a little over-powered and because it's better for environments where Denial Of Service by file attacks, are common. Added a exclusion of sym-links for the log file, this will help with the overall size of final logs. Added a option to not change the last access date of files in the compression.